+++ amazonorders/selectors.py
@@ -167,6 +167,6 @@ class Selectors:
     FIELD_TRANSACTION_ORDER_NUMBER_SELECTOR = [
         "div.apx-transactions-line-item-component-container div .a-span12"]
     FIELD_TRANSACTION_ORDER_LINK_SELECTOR = [
-        "div.apx-transactions-line-item-component-container > div:nth-child(2) a.a-link-normal"]
+        "div.apx-transactions-line-item-component-container a.a-link-normal"]
     FIELD_TRANSACTION_SELLER_NAME_SELECTOR = [
-        "div.apx-transactions-line-item-component-container > div:nth-child(3) span.a-size-base"]
+        "div.apx-transactions-line-item-component-container :has(a.a-link-normal) + div"]